Transfer code to ParamStruct
